+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* [#20](https://github.com/dblock/slack-gamebot/issues/20): Added support for matches - [@dblock](https://github.com/dblock).
 * Fix: ignore unplayed challenges during current season in `gamebot season` - [@dblock](https://github.com/dblock).
 * [#4](https://github.com/dblock/slack-gamebot/issues/4): Added support for seasons - [@dblock](https://github.com/dblock).
 * [#17](https://github.com/dblock/slack-gamebot/pull/17): Show rank section for N players - [@wrgoldstein](https://github.com/wrgoldstein).